+++ NEWS.md
@@ -1,3 +1,4 @@
+* Dots now take on colour of their lines in `sensitivity_graphs()`.
 * New default for nominal points in sensitivity graphs:
   filled black dot.
 * New option to use LaTeX for legend labels in