+++ NEWS.md
@@ -1,8 +1,9 @@
 * Dots now take on colour of their lines in `sensitivity_graphs()`.
-* New default for nominal points in sensitivity graphs:
-  filled black dot.
 * New option to use LaTeX for legend labels in
   `rebound_terms_graph()`.
+* New tests for new features.
+    - Now at 455 tests, all passing.
+    - Test coverage remains at 100%.
 
 
 # ReboundTools 0.1.14 (2021-03-14)